+++ Backend/proyectoWeb/src/main/java/com/example/demo/repositorio/AdministradoresRepositorio.java
@@ -18,8 +18,8 @@ public interface AdministradoresRepositorio extends JpaRepository<Administradore
 	public List<Administradores> findByTelefono(String telefono);
 	public Administradores findByGestiones(Gestion_Alquiler gestion);
 	public List<Administradores> findByEmail(String email);
-	Optional <Administradores> findByIdAdministrador(Integer idAdministrador);
+	Optional <Administradores> findByIdAdministrador(Long idAdministrador);
 
 	@Query(value="select password from administradores where id_administrador=:id",nativeQuery=true)
-	public String buscarAdministrador(@Param("idAdministrador")Integer idAdministrador);
+	public String buscarAdministrador(@Param("idAdministrador")Long idAdministrador);
 }
