+++ Backend/proyectoWeb/src/main/java/com/example/demo/repositorio/VehiculosRepositorio.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import com.example.demo.modelo.Usuarios;
 import com.example.demo.modelo.Vehiculos;
@@ -14,7 +15,7 @@ public interface VehiculosRepositorio extends JpaRepository<Vehiculos, Integer> 
 	
 		
 	@Query(value="update vehiculo set estado = :estado where id_vehiculo = :id",nativeQuery=true)
-	void estadoDispo(@Param("estado")String estado,@Param("id")int id_vehiculo);
+	void estadoDispo(@Param("estado")String estado,@Param("id")Long id_vehiculo);
 
 	Optional <Vehiculos> findByIdVehiculo(int idVehiculo);
 	public List<Vehiculos> findByPlaca(String placa);
